requirements.txt
# Core scanning
scapy>=2.5.0
requests>=2.31.0
paramiko>=3.4.0
nvdlib>=0.7.6
pysnmp-lextudio>=6.1.0

# Data & config
numpy>=1.24.0
pyyaml>=6.0
psutil>=5.9.0

# Web frontend
fastapi>=0.128.0
uvicorn[standard]>=0.30.0

# RL (sonraki faz icin)
gymnasium>=0.29.0

# Test & dev
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
